fix error 500 when deleting namespace (#681)

+++ libsql-server/sqld/src/namespace/mod.rs
@@ -123,7 +123,9 @@ impl MakeNamespace for PrimaryNamespaceMaker {
             }
         }
 
-        tokio::fs::remove_dir_all(ns_path).await?;
+        if ns_path.try_exists()? {
+            tokio::fs::remove_dir_all(ns_path).await?;
+        }
 
         Ok(())
     }
